# TPM/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Promo Master API
  description: |
    Trade Promotion Management System API

    ## Overview
    Promo Master is an enterprise-grade Trade Promotion Management (TPM/TPO) system
    designed for FMCG companies to manage promotions, claims, and financial operations.

    ## Authentication
    All endpoints (except `/auth/login`) require JWT authentication via Bearer token.
    Include the token in the Authorization header: `Authorization: Bearer <token>`

    ## Rate Limiting
    - General endpoints: 100 requests/minute
    - Auth endpoints: 5 requests/minute
    - Heavy operations: 10 requests/minute

    ## Error Handling
    All errors follow a standard format with error codes.
    See the Error Codes section for details.

    ## Pagination
    List endpoints support pagination via `page` and `pageSize` query parameters.
    Default: page=1, pageSize=10, max pageSize=100

  version: 2.0.0
  contact:
    name: API Support
    email: api-support@your-domain.com
  license:
    name: Proprietary
servers:
  - url: https://api.your-domain.com/api
    description: Production
  - url: https://staging-api.your-domain.com/api
    description: Staging
  - url: http://localhost:3000/api
    description: Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
...
security:
  - bearerAuth: []
paths:
...
  /claims:
    get:
      tags: [Claims]
      summary: List claims
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: promotionId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of claims

    post:
      tags: [Claims]
      summary: Create claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [promotionId, amount, type]
              properties:
                promotionId:
                  type: string
                amount:
                  type: number
                type:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Claim created
...